+++ INF404/TP5/analyse_lexicale.h
@@ -45,13 +45,13 @@
 
    typedef struct {
      char *nom;     // nom de la variable
-     int valeur;        // valeur de la variable
+     double valeur;        // valeur de la variable
    } Symbole;
 
 
    int verifier_variable(char *nom);
-   void ajouter_variable(char *nom, int val);
-   void modifier_variable(char *nom, int val);
+   void ajouter_variable(char *nom, double val);
+   void modifier_variable(char *nom, double val);
    int get_valeur_variable(char *nom);
    void affichage_table_symbole();
